requirements.txt
flask==2.3.3
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
openai>=1.0.0
python-dotenv==1.0.0
requests==2.31.0 